did more cleanup and fully fleshed out the user session workflow(s)
